Made it possible to display one reward in the enchanting table, whilst actually giving a completely different reward. This will be used when setting the desired maximum enchanting level.
If say, this should be 100 instead of 30, the plugin will display levels
in the range of 1 - 100, but in actuality, the rewards will be in the
real range of 1 - 30. This saves us from implementing the Minecraft
enchanting system from scratch, and ensures that, if it should ever
change, we don't have to start completely fresh. 

The reflection magic relies heavily on reading from the
CraftBukkit-specific classes, and also on some Minecraft-only methods.
If Minecraft is reobfuscated, the class might break. So, as a
precaution, I might attempt to extract the name from the call stack, if
the simple name lookup fails.