Ignore NULL in processClickEvent.
User Joymo reported of a NPE in StandardBlockService.processClickEvent
(ticket #4) on line 48, which, given the dumped values, occurred because
block was NULL somehow. 

While I was unable to reproduce the problem, the conditions for it to
occur (lag?) might be rare and a bigger problem than failing to handle
custom blocks (only for Tekkit though). So, we'll simply ignore the NULL
value when we get it - it's not really a problem since Bukkit has
already done most of our work for us and given the correct inventory
type. We only really need it for Tekkit custom block providers, which
should already handle the NULL case.