Don't ignore cancelled interact events.
I have a hunch that ticket #4 could be caused by a mischievous plugin
canceling the interact event in order to add custom handling. If that's
the case, we still want to record the interact event and the potential
block.

The "block" value only needs to be correct in the event the player
actually is using an inventory screen, so it doesn't matter if we don't 
respect the cancel flag. If a player clicks a workbench, and is
disallowed by a protection plugin, the worst that can happen is that we
store the block in memory. But if we miss a value, it might cause
problems for any custom block providers. The standard block provider is
unaffected by this issue as of commit
@bbc5c941f94b31a7f4b8d050ede2aa9b112bdff3.