Added the ability to set the experience drop from players.
Modified the listener such that it recognizes the player as an entity,
and changed the parser to accept "Player" as an entity (interally it
uses the entity ID -1). 

Future changes will allow the user to specify a group or a set of
players to match, and not just match every possible player.