Filtered out incorrect close inventory events.
REPRODUCING THE PROBLEM

Sometimes Bukkit would raise a close inventory event even though the
user hasn't closed any inventory screen. This problem can be reproduced
by doing the following:

 1. Place a workbench on the ground. 
 2. Position the cursor on a solid block just next to the border
    to the workbench. For instance, if the ground in front of the
    workbench is solid, move the cursor down until it just barely
    crossed the border between the workbench and the ground block.
 3. Rapidly right-click on this block as fast as possible. Do this
    for about a second or two.
 4. While right-clicking, start to move the cursor towards the
    workbench.
 5. Once the cursor reaches the workbench block, the crafting screen
    will appear. If you're using 2.4.0, this will likely lead to a 
    NullPointerException. 
    
You might have to repeat this process a couple of times until you get
the hang of it, but once you do that, you should be able to reproduce 
this problem every time. 


SOLUTION

This commit fixes resolves this problem once and for all by comparing
the time of the most recent interact event with the close inventory 
event. If this time difference is too small (under a tick or 50 ms), 
we can be sure the user wouldn't have had time to close it himself, so
it must be incorrect. 

These events are simply ignored, and will print a "Swallowed a close
event." to the debug stream.