Fixed short potion queries.
Certain potion queries would fail to load properly, if the query only
contains a type, like so:
  potion|fire resistance:
    brewing result: 12
    
This was caused by the item durability parser not preventing the text
parser from "removing" the type token, even when it would fail to parse
the durability field as an integer. Instead, it would return NULL. A
simple fix was to throw an exception instead of returning null.

In addition, added some extra debugging statements. I really had to dig
deep to solve this issue.