Adding support for parameters.
This required a massive overhaul of the underlying framework. Had to
replace every use of SampleRange with a more generic class, and every
parser has to know of the parameters that are permitted for each 
reward action. 

Even core interfaces had to be changed. This commit represents a 
BREAKING change, and we might have to bump the major version - at least
the minor version.