Switching to a "service" oriented parameter system.
Looking at the result, I realize the parameter system is extremely 
rigid and brittle - the only way to extend it is to modify the code, 
and then you have to modify it all over the place.

So, we'll switch to a more sensible "service" system where parameter 
providers register to a central repository, which is then carried over
to the listeners and parsers. This is pretty much how the custom block
providers and reward/channel providers work as well.

The only difference is that we need to support four types of
services - those that extract parameters from a player, those that
do the same for entities, and finally for blocks and items. This is
so that there can be no confusion on which object the parameter
providers are intended for.