Remove the "most specified" rule for querying elements.
To explain this rule, let us consider the following scenario:
  mobs:
    zombie: 10
    ?: 5

Which of these queries applies if we kill a zombie? Normally, it
is the last written query that will apply in case of multiple queries,
but the "most specified"-rule complicated this a bit.

The rule states that a query with more specified attributes
should always take precedent over a query with fewer (more general), 
regardless of the order they appear in. This means that the "?" will 
NOT apply, since it has no attributes, and "zombie" will override it. 

The original idea was that the order the queries appear in shouldn't
matter, that the only time order would be considered is in the case
of a tie-breaker. But this makes it impossible for someone to overwrite
or (in future commits) modify existing queries with a general query.

For instance, it would be nice if we could support a more targeted 
version of the multiplier setting. One way is to add inheritance to 
the rules (in presets):
?|world1:
  local:
    ?:
      block source:
        inherit: true
        multiplier: 5
  file:
    - config.yml
    
The "?" query would then match all "block source" queries in config.yml
and increase their yield by 400%. But with the "most specified" rule, 
this is impossible. 

So, it's better to make a breaking change now and allow for future
extensibility. Though, there are problems with this approach too. For
instance, if you combine these two configuration files:

> config.yml
mobs:
  ?: 0
  zombie: 5
  
> config2.yml
mobs:
  ?: 0
  creeper: 5

> presets.yml
?:
  file:
    - config.yml
    - config2.yml
    
Then the "?" rule in config2 will overwrite every value in "config". 
But it is possible to work around this limitation, for instance by
moving general queries to a separate configuration file. The "most
specified" rule doesn't allow for any workarounds.

It might be possible to add a way to specify whether or not to use the
"most specified" rule in the future, but I would have to determine
if there's any demand for it. Probably not, to be honest.